+++ docs/source/changelog.md
@@ -34,6 +34,8 @@ See [build docs](building-pyzmq) for more info.
 __New__:
 
 - Experimental support for wheels on windows-arm64
+- `Socket.bind('tcp://ip:0')` can be used as a context manager to bind to a random port.
+  The resulting URL can be retrieved as `socket.last_endpoint`.
 
 __Enhancements__: